asynchronous client side js added

+++ controllers/admin.js
@@ -1,6 +1,6 @@
 const { validationResult } = require("express-validator");
 const mongoose = require("mongoose");
-
+const fileHelper = require("../util/file");
 const Product = require("../models/product");
 
 exports.getAddProduct = (req, res, next) => {
@@ -19,8 +19,8 @@ exports.postAddProduct = (req, res, next) => {
 	const image = req.file;
 	const price = req.body.price;
 	const description = req.body.description;
-  if(!image){
-    return res.status(422).render("admin/edit-product", {
+	if (!image) {
+		return res.status(422).render("admin/edit-product", {
 			pageTitle: "Add Product",
 			path: "/admin/add-product",
 			editing: false,
@@ -30,10 +30,10 @@ exports.postAddProduct = (req, res, next) => {
 				price: price,
 				description: description,
 			},
-			errorMessage: 'File format not supported',
-			validationErrors: []
+			errorMessage: "File format not supported",
+			validationErrors: [],
 		});
-  }
+	}
 	console.log(image);
 	const errors = validationResult(req);
 
@@ -47,14 +47,14 @@ exports.postAddProduct = (req, res, next) => {
 			product: {
 				title: title,
 				price: price,
-        imageUrl:image,
+				imageUrl: image,
 				description: description,
 			},
 			errorMessage: errors.array()[0].msg,
 			validationErrors: errors.array(),
 		});
 	}
-const imageUrl=image.path;
+	const imageUrl = image.path;
 
 	const product = new Product({
 		title: title,
@@ -153,9 +153,11 @@ exports.postEditProduct = (req, res, next) => {
 			product.title = updatedTitle;
 			product.price = updatedPrice;
 			product.description = updatedDesc;
-      if(image){
-        product.imageUrl = image.path;
-      }
+			if (image) {
+
+				fileHelper.deleteFile(product.imageUrl);
+				product.imageUrl = image.path;
+			}
 			return product.save().then((result) => {
 				console.log("UPDATED PRODUCT!");
 				res.redirect("/admin/products");
@@ -173,7 +175,7 @@ exports.getProducts = (req, res, next) => {
 		// .select('title price -_id')
 		// .populate('userId', 'name')
 		.then((products) => {
-			console.log(products);
+			// console.log(products);
 			res.render("admin/products", {
 				prods: products,
 				pageTitle: "Admin Products",
@@ -187,16 +189,21 @@ exports.getProducts = (req, res, next) => {
 		});
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-	const prodId = req.body.productId;
-	Product.deleteOne({ _id: prodId, userId: req.user._id })
+exports.deleteProduct = (req, res, next) => {
+	const prodId = req.params.productId;
+	Product.findById(prodId)
+		.then((product) => {
+			if (!product) {
+				return next(new Error("Product not Found"));
+			}
+			fileHelper.deleteFile(product.imageUrl);
+			return Product.deleteOne({ _id: prodId, userId: req.user._id });
+		})
 		.then(() => {
 			console.log("DESTROYED PRODUCT");
-			res.redirect("/admin/products");
+			res.status(200).json({message:'Success'});
 		})
 		.catch((err) => {
-			const error = new Error(err);
-			error.httpStatusCode = 500;
-			return next(error);
+			res.status(500).json({message:'Deleting Product failed'});
 		});
 };
